Move utils::printBuildFlags into its own header 'build_info.h'
Reduce unnecessary dependencies to caps if they are not needed, e.g., in
basic unit tests.

Note: By including the header the actual build flags for the current
      compilation unit will be shown instead of the once applying only
      to 'utils.cpp'.
