+++ src/utils.cpp
@@ -173,26 +173,6 @@ namespace utils {
 		}
 	}
 
-
-	// helper to print flags
-	void printBuildFlags() {
-		bool parallel = false;
-		bool userdebug = false;
-		bool capa = false;
-
-#ifdef WS_PARALLEL
-		parallel = true;
-#endif
-#ifdef WS_ALLOW_USER_DEBUG
-		userdebug = true;
-#endif
-#ifdef WS_CAPA
-		capa = true;
-#endif
-		fmt::println("Build flags: WS_PARALLEL={}, WS_CAPA={}, WS_ALLOW_USER_DEBUG={}",parallel,capa,userdebug);
-		fmt::println("Runtime flags: isusermode={}, issetuid={}, hascaps={}", caps.isUserMode(), caps.isSetuid(), caps.hasCaps());
-	}
-
 	// we only support C locale, if the used local is not installed on the system
 	// ws_allocate fails, this should be called in all tools
 	//  problem showed with a remote SuSE machine with DE locale, coming through ssh
