+++ src/ws_list.cpp
@@ -52,6 +52,7 @@
 #include <boost/program_options.hpp>
 #include "config.h"
 
+#include "build_info.h"
 #include "db.h"
 #include "user.h"
 #include "fmt/base.h"
