+++ tests/UserConfig_test.cpp
@@ -1,15 +1,11 @@
-#include "../src/UserConfig.h"
-
 #define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
 #include <catch2/catch_test_macros.hpp>
 
 #include "../src/UserConfig.h"
-#include "../src/caps.h"
 
 bool debugflag = false;
 bool traceflag = false;
 
-Cap caps{};
 
 TEST_CASE("User config yaml format", "[userconfig]") {
